Fixed warnings when compiling with MSYS.
git-svn-id: http://crystal.svn.sourceforge.net/svnroot/crystal/CS/branches/soc2011/selfshadow@36570 8cc4aa7f-3514-0410-904f-f2cc9021211c

+++ plugins/rendermanager/osm/shadow_osm.h
@@ -213,7 +213,7 @@ namespace CS
             CS_ASSERT_MSG("Null node encountered, should not be possible", node);
 
             // setup split dists
-            for (int i = 0 ; i < node->meshes.GetSize(); i ++)
+            for (size_t i = 0 ; i < node->meshes.GetSize(); i ++)
             {
               typename RenderTree::ContextNode::TreeType::MeshNode::SingleMesh 
                 mesh = node->meshes.Get(i);
